Scrape with product specific URLs
Previously, the web scraper would scrape the supermarket
website based on "category" URLs. This would mean the
application would have to loop through all products in
the category until the correct product was found. This
was problematic in the case of a category whose total
results exceeded the amount of results a single page
could present by default. Meaning that if the wanted
product was on the second page of results, the scraper
could not find it. Therefore it was changed to work with
product specific URLs, making it more effective and
efficient.