Download products.json from remote repo (#3)
The products.json file is now located in a private
GitHub repo which the scraper downloads automatically
at the beggining of execution. This prevents others
from seeing private product information and allows
any user to simply edit that file and the scraper
will take care of the rest.

README.md file needs to be updated next to reflect
these changes.